+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Make `do generate` supports system locale.
+- Add `"svg"` as a output format for `do <INPUT> typst-compile`.
 
 ### Changed